requirements.txt
httpx>=0.21.1
crontab~=1.0.1
PyYAML~=6.0
beautifulsoup4
PyYAML